fix: add fetchPosts component to take care of posts, instead rerender whole page on refetch posts
